Features/#18/create web demo (#20)
* trying to get graphics to images

* testing checks

* working demo

* working on API docs

* to push

* layout navbar

* API REFERFENCES Label / header value

* template for docs

* Add a demo controller

* working API

* Finalizing and ready for publish

* Fixing Code factor errors

* Fix codefactor formatting issues

+++ src/TalkingCSharp.PrintityFramework/PFW_TableColumn.cs
@@ -63,7 +63,7 @@ public class PFW_TableColumn : IPFW_DrawableColumn
         return this;
     }
 
-    public PFW_TableColumn SetWidth(int width, PFW_MeasurementsEnum unit)
+    public PFW_TableColumn SetWidth(float width, PFW_MeasurementsEnum unit)
     {
         this.Width = width;
         this.WidthUnit = unit;
